[2.0.0] Initial Release  - Updated to Ubuntu 16.0.4 (ubuntu/xenial64)  - Removed config_*.sh files from root and made it so all configuration is done via config.yml  - Updated README.md with new software versions  - Changes to Vagrantfile for config_*.sh generation, optional software setup, removed unneeded second declaration of "dir" variable  - Added optional software setup/config for ElasticSearch/RabbitMq (Solr still to come)  - Updates to scripts/200-setup_varnish.sh for Varnish 4.x  - Updates to scripts/400-setup_mysql.sh for Percona 5.6 build differences on 16.0.4  - Added mod proxy_balancer to scripts/100-setup_apache.sh  - Updated PHP init.d template to use unique "Provides:" header  - Updated vhost.sh to use SetHandler for php instead of ProxyPassMatch  - Added pre-built archives of php versions  - Updated scripts/500-setup_php.sh to now use the included pre-built php archives instead of building from source when available

+++ scripts/500-setup_php.sh
@@ -118,12 +118,18 @@ function setup_php() {
         phpv=${arr[0]}
         phpn=${arr[1]}
         phpp=${arr[2]}
+        phpb=${arr[3]}
 
         if [ ! -f /opt/phpfarm/inst/php-${phpv}/bin/php ]; then
-            cd /opt/phpfarm/src
-            ./main.sh ${phpv}
-            setup_xdebug ${phpv}
-            cp /opt/phpfarm/inst/php-${phpv}/etc/php.ini /opt/phpfarm/inst/php-${phpv}/lib/php.ini
+            if [ ${phpb} == 'true' ] || [ ! -f /vagrant/files/php_builds/php-${phpv}.tar.gz ] ; then
+                cd /opt/phpfarm/src
+                ./main.sh ${phpv}
+                setup_xdebug ${phpv}
+                cp /opt/phpfarm/inst/php-${phpv}/etc/php.ini /opt/phpfarm/inst/php-${phpv}/lib/php.ini
+             else
+                cd /opt/phpfarm/inst
+                tar -zxf /vagrant/files/php_builds/php-${phpv}.tar.gz
+             fi
         fi
         if [ ${phpv:0:1} == 5 ]; then
             php_config_suffix_escaped="\/etc\/php-fpm.conf"
